:recycle: refactor: Nav 컴포넌트 분리 및 routes 폴더 변경
