+++ craco.config.js
@@ -9,7 +9,6 @@ module.exports = {
       '@api': path.resolve(__dirname, 'src/api'),
       '@utils': path.resolve(__dirname, 'src/utils'),
       '@styles': path.resolve(__dirname, 'src/styles'),
-      '@router': path.resolve(__dirname, 'src/router'),
     },
   },
   jest: {
@@ -20,7 +19,6 @@ module.exports = {
         '^\\@api/(.*)$': '<rootDir>/src/api/$1',
         '^\\@utils/(.*)$': '<rootDir>/src/utils/$1',
         '^\\@styles/(.*)$': '<rootDir>/src/styles/$1',
-        '^\\@router/(.*)$': '<rootDir>/src/router/$1',
       },
     },
   },
