+++ src/App.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import Router from '@router/Router';
+import Router from './Router';
 
 const App = () => {
-  return (<Router />);
+  return <Router />;
 };
 
 export default App;
