+++ src/components/index.jsx
@@ -1,3 +1,3 @@
 export { default as CurrentCurrency } from './CurrentCurrency';
 export { default as SelectCurrency } from './SelectCurrency';
-
+export { default as Nav } from './Nav';
